+++ CHANGES.md
@@ -1,4 +1,11 @@
 
+v.0.10.0
+
+- Add Decimal type for handling decimal numbers correctly
+- Add IsDecimal validator
+- Breaking change to how the results are returned. When using the Chain only keys with errors are returned now.
+- Fix IsLength failing when passed ValueNotSet()
+
 v.0.9.1
 
 - Fix IsRange failing when passed ValueNotSet()